# engine/docs2/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flapjack API",
    "description": "Drop-in replacement for Algolia search. Provides full-text search with filters, facets, geo-search, and more.",
    "contact": {
      "name": "Flapjack",
      "url": "https://github.com/griddlehq/flapjack"
    },
    "license": {
      "name": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:7700",
      "description": "Local development"
    },
    {
      "url": "https://fj-us-west-1.flapjack.foo",
      "description": "Production US West"
    }
  ],
  "paths": {
...
    "/1/indexes/{indexName}/queries": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Execute multiple search queries in a single request, optionally merging results via federation.",
        "operationId": "batch_search",
        "parameters": [
          {
            "name": "indexName",
            "in": "path",
            "description": "Index to search",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Batch search request with multiple queries",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Batch search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchSearchResponse"
                }
              }
            }
          },
          "404": {
            "description": "Index not found"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchSearchLegacyResponse": {
        "type": "object",
        "required": [
          "results"
        ],
        "properties": {
          "results": {
            "type": "array",
            "items": {}
          }
        }
      },
      "BatchSearchRequest": {
        "type": "object",
        "description": "Batch search request supporting legacy parallel queries plus federated merge options.",
        "required": [
          "requests"
        ],
        "properties": {
          "federation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FederationConfig"
              }
            ]
          },
          "requests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchRequest"
            }
          },
          "strategy": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "BatchSearchResponse": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/BatchSearchLegacyResponse"
          },
          {
            "$ref": "#/components/schemas/FederatedResponse"
          }
        ]
      },
...
      "FederatedHit": {
        "type": "object",
        "description": "A federated hit preserves arbitrary document fields while making `_federation`\nmetadata explicit in both runtime values and the published schema.",
        "required": [
          "_federation"
        ],
        "properties": {
          "_federation": {
            "$ref": "#/components/schemas/FederationMeta"
          }
        },
        "additionalProperties": {}
      },
      "FederatedResponse": {
        "type": "object",
        "description": "Flat federated batch response returned when the top-level `federation` field is present.",
        "required": [
          "hits",
          "estimatedTotalHits",
          "limit",
          "offset"
        ],
        "properties": {
          "estimatedTotalHits": {
            "type": "integer",
            "minimum": 0
          },
          "hits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FederatedHit"
            }
          },
          "limit": {
            "type": "integer",
            "minimum": 0
          },
          "offset": {
            "type": "integer",
            "minimum": 0
          },
          "processingTimeMS": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "FederationConfig": {
        "type": "object",
        "description": "Top-level federated batch settings shared by request parsing and OpenAPI.",
        "properties": {
          "limit": {
            "type": "integer",
            "minimum": 0
          },
          "mergeFacets": {
            "type": [
              "object",
              "null"
            ]
          },
          "offset": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "FederationOptions": {
        "type": "object",
        "properties": {
          "weight": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "HybridSearchParams": {
        "type": "object",
        "description": "Parameters for hybrid (keyword + vector) search.\n\nMeilisearch-style: `\"hybrid\": {\"semanticRatio\": 0.8, \"embedder\": \"mymodel\"}`\nAlgolia-style: synthesized internally when `mode: \"neuralSearch\"`.",
        "properties": {
          "embedder": {
            "type": "string"
          },
          "semanticRatio": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "IgnorePluralsValue": {
        "oneOf": [
          {
            "type": "string",
            "enum": [
              "Disabled"
            ]
          },
          {
            "type": "string",
            "enum": [
              "All"
            ]
          },
          {
            "type": "object",
            "required": [
              "Languages"
            ],
            "properties": {
              "Languages": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        ]
      },
      "IndexMode": {
        "type": "string",
        "enum": [
          "keywordSearch",
          "neuralSearch"
        ]
      },
...
      "RemoveStopWordsValue": {
        "oneOf": [
          {
            "type": "string",
            "enum": [
              "Disabled"
            ]
          },
          {
            "type": "string",
            "enum": [
              "All"
            ]
          },
          {
            "type": "object",
            "required": [
              "Languages"
            ],
            "properties": {
              "Languages": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        ]
      },
...
      "SearchRequest": {
        "type": "object",
        "description": "Algolia-compatible search request supporting both JSON body fields and URL-encoded `params`.\n\nCovers keyword search, facet lookups, geo filtering, analytics flags, hybrid\nsearch controls, and federated batch-only options such as per-query weights.",
        "properties": {
          "advancedSyntax": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "advancedSyntaxFeatures": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "alternativesAsExact": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "analytics": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "analyticsTags": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "aroundLatLng": {
            "type": [
              "string",
              "null"
            ]
          },
          "aroundLatLngViaIP": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "aroundPrecision": {},
          "aroundRadius": {},
          "attributesToHighlight": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "attributesToRetrieve": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "attributesToSnippet": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "clickAnalytics": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "decompoundQuery": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Whether to split compound words in Germanic languages (de, nl, fi, da, sv, no).\nDefault: true (matching Algolia behavior). Set to false to disable."
          },
          "disableExactOnAttributes": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "distinct": {},
          "enableABTest": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "enablePersonalization": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "enableReRanking": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "enableRules": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "enableSynonyms": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "exactOnSingleWordQuery": {
            "type": [
              "string",
              "null"
            ]
          },
          "facet": {
            "type": [
              "string",
              "null"
            ],
            "description": "Facet name for type=facet multi-search queries"
          },
          "facetFilters": {},
          "facetQuery": {
            "type": [
              "string",
              "null"
            ],
            "description": "Facet query string for type=facet multi-search queries"
          },
          "facetingAfterDistinct": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "facets": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "federationOptions": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FederationOptions"
              }
            ]
          },
          "filters": {
            "type": [
              "string",
              "null"
            ]
          },
          "getRankingInfo": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "highlightPostTag": {
            "type": [
              "string",
              "null"
            ]
          },
          "highlightPreTag": {
            "type": [
              "string",
              "null"
            ]
          },
          "hitsPerPage": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "hybrid": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/HybridSearchParams"
              }
            ]
          },
          "ignorePlurals": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/IgnorePluralsValue"
              }
            ]
          },
          "indexName": {
            "type": [
              "string",
              "null"
            ]
          },
          "insideBoundingBox": {},
          "insidePolygon": {},
          "maxFacetHits": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Max facet hits for type=facet multi-search queries",
            "minimum": 0
          },
          "maxValuesPerFacet": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "minProximity": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "minimumAroundRadius": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "mode": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/IndexMode"
              }
            ]
          },
          "naturalLanguages": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "numericFilters": {},
          "optionalFilters": {},
          "page": {
            "type": "integer",
            "minimum": 0
          },
          "params": {
            "type": [
              "string",
              "null"
            ],
            "description": "URL-encoded params string (used by multi-query). Merged during deserialization."
          },
          "percentileComputation": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "personalizationFilters": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "personalizationImpact": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "query": {
            "type": "string"
          },
          "queryLanguages": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "queryType": {
            "type": [
              "string",
              "null"
            ]
          },
          "reRankingApplyFilter": {
            "type": [
              "string",
              "null"
            ]
          },
          "relevancyStrictness": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "removeStopWords": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RemoveStopWordsValue"
              }
            ]
          },
          "removeWordsIfNoResults": {
            "type": [
              "string",
              "null"
            ]
          },
          "replaceSynonymsInHighlight": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "responseFields": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "restrictHighlightAndSnippetArrays": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "restrictSearchableAttributes": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "ruleContexts": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "sessionID": {
            "type": [
              "string",
              "null"
            ]
          },
          "similarQuery": {
            "type": [
              "string",
              "null"
            ]
          },
          "snippetEllipsisText": {
            "type": [
              "string",
              "null"
            ]
          },
          "sort": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "sortFacetValuesBy": {
            "type": [
              "string",
              "null"
            ]
          },
          "sumOrFiltersScores": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "tagFilters": {},
          "type": {
            "type": [
              "string",
              "null"
            ]
          },
          "typoTolerance": {},
          "userToken": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-algolia-api-key"
      }
    }
  },
...
}